refactor: provider tracking code
